refatorando
