+++ pokedex/src/App.js
@@ -1,7 +1,5 @@
 import './App.css';
 import Router from "./routes/Routes";
-import { ThemeProvider } from 'styled-components'
-import {theme} from './theme/theme'
 import { GlobalStyle } from './styles';
 import GlobalState from './GlobalStateContext/GlobalState';
 
@@ -11,10 +9,8 @@ function App() {
   return (
     <div>
       <GlobalState>
-        <ThemeProvider theme={theme}>
           <GlobalStyle />
           <Router/>
-        </ThemeProvider>
       </GlobalState>
     </div>
   );
