+++ pokedex/src/components/headers/headerHome/HeaderHome.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import {useNavigate} from "react-router-dom";
 import styled from 'styled-components';
-import { goToPokedex} from "../../coordinator/Coordinator"
-import { FluxoryButtons, Header, Title } from '../../styles';
+import { goToPokedex} from "../../../coordinator/Coordinator"
+import { FluxoryButtons, Header, Title } from '../../../styles';
 
 const ButtonDiv = styled.div`
 justify-self: end;
@@ -12,9 +12,7 @@ const HeaderHome = () => {
     const navigate = useNavigate()
   return (
     <Header>
-      {/* <div> */}
         <Title>Lista de Pokémons</Title>
-      {/* </div> */}
       <ButtonDiv>
         <FluxoryButtons onClick={() => goToPokedex(navigate)}>Ir para pokedex</FluxoryButtons>
       </ButtonDiv>
