+++ pokedex/src/pages/pokedex/Pokedex.js
@@ -1,8 +1,5 @@
 import React, { useContext } from 'react'
-import HeaderPokedex from '../../components/headerPokedex/HeaderPokedex'
-import { goToPokemonDetail } from '../../coordinator/Coordinator'
-import { useNavigate } from "react-router-dom";
-import { FluxoryButtons } from '../../styles';
+import HeaderPokedex from '../../components/headers/headerPokedex/HeaderPokedex'
 import { PokemonsContext } from '../../GlobalStateContext/Context';
 import PokeCard from '../../components/cardPokemon/PokeCard';
 import { DivCard } from '../home/Home';
