--- pokedex/src/theme/theme.js
@@ -1,6 +1,0 @@
-export const theme = {
-    titulo: 'orange',
-    bg: '#fff',
-    titles: 'blue',
-}
-
